+++ out/tags.js
@@ -5,6 +5,7 @@ exports.ScanDocumentForADSCompilerTag = ScanDocumentForADSCompilerTag;
 const vscode = require("vscode");
 const application = require("./application");
 const configuration = require("./configuration");
+const filesystem = require("./filesystem");
 function ScanDocumentForADSLanguageTag(document) {
     // prepare
     const text = document.getText();
@@ -26,16 +27,17 @@ function ScanDocumentForADSLanguageTag(document) {
         });
     }
     ;
-    // Is in the samples folder? if so lets force it based on the subfolder
-    const documentFileName = document.fileName.toLowerCase();
-    if (documentFileName.includes("samples")) {
+    // Is in the samples folder and a language document? if so determine which language to use
+    const fileName = document.fileName.toLowerCase();
+    const fileExtension = filesystem.GetFileExtension(document.uri).toLowerCase();
+    if (fileName.includes("samples") && fileExtension == ".bas") {
         // Prepare
         let matchingLanguage = '';
         // Validate
-        if (documentFileName.includes(application.SeventyEightHundredBasicLanguageId)) {
+        if (fileName.includes(application.SeventyEightHundredBasicLanguageId)) {
             matchingLanguage = application.SeventyEightHundredBasicLanguageId;
         }
-        else if (documentFileName.includes(application.BatariBasicLanguageId)) {
+        else if (fileName.includes(application.BatariBasicLanguageId)) {
             matchingLanguage = application.BatariBasicLanguageId;
         }
         // is the language available?