+++ src/extension.ts
@@ -8,7 +8,7 @@ import * as tags from "./tags";
 import * as configuration from "./configuration";
 import { WelcomePage } from './pages/welcome';
 import { SpriteEditorPage } from './pages/spriteeditor';
-import { SamplesPage } from './pages/samples';
+import { LearningCenterPage } from './pages/learningcenter';
 import './statusbar';
 
 // Activation Events
@@ -23,7 +23,7 @@ export async function activate(context: vscode.ExtensionContext) {
 	// Pages
 	let welcomePage = new WelcomePage();
 	let spriteEditorPage = new SpriteEditorPage();
-	let samplesPage = new SamplesPage();
+	let learningCenterPage = new LearningCenterPage();
 
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
@@ -47,10 +47,10 @@ export async function activate(context: vscode.ExtensionContext) {
 		console.log('User activated command "extension.openSpriteEditorPage"');
 		spriteEditorPage.openPage(context);
 	});
-	// Samples
-	const openSamplesPage = vscode.commands.registerCommand('extension.openSamplesPage', () => {
-		console.log('User activated command "extension.openSamplesPage"');
-		samplesPage.openPage(context);
+	// LearningCenter
+	const openLearningCenterPage = vscode.commands.registerCommand('extension.openLearningCenterPage', () => {
+		console.log('User activated command "extension.openLearningCenterPage"');
+		learningCenterPage.openPage(context);
 	});
 	// 2600 editors
 	const openPlayerPalPage = vscode.commands.registerCommand('extension.openPlayerPalPage', () => {